requirements.txt
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
redis>=4.5.0
websockets>=11.0.0
prometheus-client>=0.16.0
tiktoken>=0.4.0
tenacity>=8.2.0
httpx>=0.24.0
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
